Additional C4244 warning fixes for changes in UE main #jira nojira #rnx #rb none
[CL 11057640 by Michael Noland in Main branch]

+++ Engine/Source/Runtime/Core/Private/Tests/Algo/UniqueTest.cpp
@@ -40,7 +40,7 @@ bool FUniqueTest::RunTest(const FString& Parameters)
 	}
 	{
 		int32 Array[] = {1};
-		int32 NewSize = Unique(Array);
+		int32 NewSize = (int32)Unique(Array);
 		TestEqual(TEXT("`Unique` must support C arrays"), NewSize, 1);
 	}
 	{
